Fixed issues seen by @akien-mga

+++ source/spells/SpellCreatureExplosion.cpp
@@ -170,7 +170,7 @@ bool SpellCreatureExplosion::castSpell(GameMap* gameMap, Player* player, ODPacke
 
     int32_t pricePerTarget = ConfigManager::getSingleton().getSpellConfigInt32("CreatureExplosionPrice");
     int32_t playerMana = static_cast<int32_t>(player->getSeat()->getMana());
-    uint32_t nbTargets = std::min(static_cast<uint32_t>(playerMana / pricePerTarget), creatures.size());
+    uint32_t nbTargets = std::min(static_cast<uint32_t>(playerMana / pricePerTarget), static_cast<uint32_t>(creatures.size()));
     int32_t priceTotal = nbTargets * pricePerTarget;
 
     if(creatures.size() > nbTargets)
